Move tests to own file
